+++ src/components/List/UserBoard/items/StatusControllerOpenToggle.tsx
@@ -1,0 +1,22 @@
+import { ReactNode, useState } from "react";
+import StatusToggleList from "./StatusToggleList";
+
+const StatusToggle = ({ children, id }: { children: ReactNode, id: number }) => {
+    const [isOpenStatus, setIsOpenStatus] = useState(false);
+    const onStatusClick = () => {
+        setIsOpenStatus((prev) => !prev);
+    }
+
+    return (
+        <div className="relative">
+            <button onClick={onStatusClick}>
+                {children}
+            </button>
+            {
+                isOpenStatus && <StatusToggleList id={id} closeList={onStatusClick} />
+            }
+        </div>
+    )
+};
+
+export default StatusToggle;
